+++ playground/Coalesce.Web.Vue3/Pages/TestPage.cshtml.cs
@@ -1,0 +1,10 @@
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace Coalesce.Web.Vue.Pages;
+
+public class TestPageModel : PageModel
+{
+    public void OnGet()
+    {
+    }
+}